+++ cmd/cycloid/middleware/middleware.go
@@ -42,7 +42,7 @@ type Middleware interface {
 	Login(email, password string) (*models.UserSession, error)
 
 	// LoginOrg is the used to log the user into a Cycloid organization
-	LoginOrg(org, child, email, password string) (*models.UserSession, error)
+	LoginOrg(org, child string) (*models.UserSession, error)
 
 	DeleteMember(org string, name string) error
 	GetMember(org string, name string) (*models.MemberOrg, error)
@@ -68,6 +68,7 @@ type Middleware interface {
 	UnpausePipelineJob(org, project, env, job string) error
 	UnpausePipeline(org string, project string, env string) error
 	UpdatePipeline(org string, project string, env string, pipeline string, variables string) (*models.Pipeline, error)
+	ListPipelines(org string) ([]*models.Pipeline, error)
 
 	CreateProject(org, projectName, projectCanonical, env, pipelineTemplate, variables, description, cloudProvider, stackRef, usecase string, configRepo uint32) (*models.Project, error)
 	DeleteProjectEnv(org, project, env string) error
@@ -82,6 +83,19 @@ type Middleware interface {
 
 	GetStack(org, ref string) (*models.ServiceCatalog, error)
 	ListStacks(org string) ([]*models.ServiceCatalog, error)
+
+	// API keys method
+	// CreateAPIKey will request API to generate and return an API key
+	CreateAPIKey(org, name, canonical, description string, roleID uint32) (*models.APIKey, error)
+
+	// ListAPIKey will request API to list generated API keys
+	ListAPIKey(org string) ([]*models.APIKey, error)
+
+	// GetAPIKey will request API to get a specified generated API key by its canonical
+	GetAPIKey(org, canonical string) (*models.APIKey, error)
+
+	// DeleteAPIKey will request API to delete a specified generated API key
+	DeleteAPIKey(org, canonical string) error
 }
 
 type middleware struct {